fix: move dto to http module
