+++ piikii-application/src/main/kotlin/com/piikii/application/domain/course/CourseService.kt
@@ -1,7 +1,6 @@
 package com.piikii.application.domain.course
 
 import com.piikii.application.port.input.course.CourseUseCase
-import com.piikii.application.port.input.course.dto.response.CourseExistenceResponse
 import com.piikii.application.port.output.persistence.CourseQueryPort
 import org.springframework.stereotype.Service
 import java.util.UUID
@@ -10,9 +9,7 @@ import java.util.UUID
 class CourseService(
     private val courseQueryPort: CourseQueryPort,
 ) : CourseUseCase {
-    override fun readCourseExistenceInRoom(roomId: UUID): CourseExistenceResponse {
-        return CourseExistenceResponse(
-            courseQueryPort.isExistCourse(roomId),
-        )
+    override fun isCourseExist(roomId: UUID): Boolean {
+        return courseQueryPort.isCourseExist(roomId)
     }
 }
