+++ piikii-application/src/main/kotlin/com/piikii/application/port/input/course/CourseUseCase.kt
@@ -1,8 +1,7 @@
 package com.piikii.application.port.input.course
 
-import com.piikii.application.port.input.course.dto.response.CourseExistenceResponse
 import java.util.UUID
 
 interface CourseUseCase {
-    fun readCourseExistenceInRoom(roomId: UUID): CourseExistenceResponse
+    fun isCourseExist(roomId: UUID): Boolean
 }
